restructure serializers
