+++ src/serializers/list_tuple.rs
@@ -11,7 +11,8 @@ use serde::ser::SerializeSeq;
 use crate::build_tools::SchemaDict;
 
 use super::any::{fallback_serialize, fallback_to_python, fallback_to_python_json, AnySerializer};
-use super::{py_err_se_err, BuildSerializer, CombinedSerializer, Extra, PydanticSerializer, TypeSerializer};
+use super::shared::{py_err_se_err, BuildSerializer, CombinedSerializer, Extra, TypeSerializer};
+use super::PydanticSerializer;
 
 type IncEx = Option<IntMap<usize, Option<PyObject>>>;
